+++ SleepModel/HEMAlarmExpansionListCell.h
@@ -8,10 +8,16 @@
 
 #import "HEMAlarmListCell.h"
 
+extern CGFloat const kHEMAlarmExpansionViewHeight;
+
 @interface HEMAlarmExpansionListCell : HEMAlarmListCell
 
 @property (weak, nonatomic) IBOutlet UIView* expansionSeparator;
 @property (weak, nonatomic) IBOutlet UIImageView* expansionIconView;
 @property (weak, nonatomic) IBOutlet UILabel* expansionLabel;
+
+- (void)showExpansionWithIcon:(UIImage*)icon
+                         text:(NSAttributedString*)attributedText
+                         tyep:(NSUInteger)type;
 
 @end