+++ SleepModel/HEMAlarmListPresenter.m
@@ -8,6 +8,8 @@
 #import <AttributedMarkdown/markdown_peg.h>
 #import <SenseKit/SENLocalPreferences.h>
 
+#import "NSMutableAttributedString+HEMFormat.h"
+
 #import "HEMAlarmListPresenter.h"
 #import "HEMSubNavigationView.h"
 #import "HEMAlarmService.h"
@@ -24,12 +26,12 @@
 #import "HEMActivityIndicatorView.h"
 #import "HEMAlarmCache.h"
 #import "HEMAlarmExpansionListCell.h"
+#import "HEMExpansionService.h"
 
 static CGFloat const HEMAlarmListButtonMinimumScale = 0.95f;
 static CGFloat const HEMAlarmListButtonMaximumScale = 1.2f;
 static CGFloat const HEMAlarmLoadAnimeDuration = 0.5f;
 static CGFloat const HEMAlarmListCellHeight = 96.f;
-static CGFloat const HEMAlarmExpansionListCellHeight = 138.0f;
 static CGFloat const HEMAlarmListNoAlarmCellBaseHeight = 292.0f;
 static CGFloat const HEMAlarmListItemSpacing = 8.f;
 static CGFloat const HEMAlarmNoAlarmHorzMargin = 40.0f;
@@ -51,15 +53,18 @@ static NSString *const HEMAlarmListTimeKey = @"alarms.alarm.meridiem.%@";
 @property (nonatomic, strong) NSDateFormatter *hour24Formatter;
 @property (nonatomic, strong) NSDateFormatter *hour12Formatter;
 @property (nonatomic, strong) NSDateFormatter *meridiemFormatter;
+@property (nonatomic, strong) HEMExpansionService* expansionService;
 
 @end
 
 @implementation HEMAlarmListPresenter
 
-- (instancetype)initWithAlarmService:(HEMAlarmService*)alarmService {
+- (instancetype)initWithAlarmService:(HEMAlarmService*)alarmService
+                    expansionService:(HEMExpansionService*)expansionService {
     self = [super init];
     if (self) {
         _alarmService = alarmService;
+        _expansionService = expansionService;
         _hour12Formatter = [NSDateFormatter new];
         _hour12Formatter.dateFormat = @"hh:mm";
         _hour24Formatter = [NSDateFormatter new];
@@ -293,7 +298,7 @@ static NSString *const HEMAlarmListTimeKey = @"alarms.alarm.meridiem.%@";
                   cellForItemAtIndexPath:(NSIndexPath *)indexPath {
     if ([[[self alarmService] alarms] count] > 0) {
         SENAlarm* alarm = [[self alarmService] alarms][[indexPath row]];
-        if ([[self alarmService] hasLightsEnabledForAlarm:alarm]) {
+        if ([[self alarmService] numberOfEnabledExpansionsIn:alarm]) {
             return [self collectionView:collectionView alarmExpansionCellAtIndexPath:indexPath];
         } else {
             return [self collectionView:collectionView alarmCellAtIndexPath:indexPath];
@@ -302,6 +307,60 @@ static NSString *const HEMAlarmListTimeKey = @"alarms.alarm.meridiem.%@";
         return [self collectionView:collectionView statusCellAtIndexPath:indexPath];
     } else {
         return [self collectionView:collectionView emptyCellAtIndexPath:indexPath];
+    }
+}
+
+- (NSAttributedString*)attributedExpansionTextFor:(SENAlarmExpansion*)expansion
+                                         forAlarm:(SENAlarm*)alarm {
+    SENExpansionValueRange range = [expansion targetRange];
+    UIColor* titleColor = [alarm isOn] ? [UIColor grey6] : [UIColor grey4];
+    NSDictionary* titleAttributes = @{NSFontAttributeName : [UIFont h7],
+                                      NSForegroundColorAttributeName : titleColor};
+    NSDictionary* valueAttributes = @{NSFontAttributeName : [UIFont h7],
+                                      NSForegroundColorAttributeName : [UIColor grey4]};
+    NSString* format = nil;
+    NSArray* args = nil;
+    
+    switch ([expansion type]) {
+        default:
+        case SENExpansionTypeLights: {
+            NSString* title = NSLocalizedString(@"alarm.light.title", nil);
+            NSString* unitSymbol = NSLocalizedString(@"measurement.percentage.unit", nil);
+            NSString* value = [NSString stringWithFormat:@"%.0f%@", range.max, unitSymbol];
+            NSAttributedString* attributedTitle = [[NSAttributedString alloc] initWithString:title attributes:titleAttributes];
+            NSAttributedString* attributedValue = [[NSAttributedString alloc] initWithString:value attributes:valueAttributes];
+            
+            args = @[attributedTitle, attributedValue];
+            format = NSLocalizedString(@"alarm.light.attribution.format", nil);
+            break;
+        }
+        case SENExpansionTypeThermostat: {
+            range = [[self expansionService] convertThermostatRangeBasedOnPreference:range];
+            NSString* title = NSLocalizedString(@"alarm.thermostat.title", nil);
+            NSString* minValue = [NSString stringWithFormat:@"%.0f", range.min];
+            NSString* maxValue = [NSString stringWithFormat:@"%.0f", range.max];
+            NSAttributedString* attributedTitle = [[NSAttributedString alloc] initWithString:title attributes:titleAttributes];
+            NSAttributedString* attributedMinValue = [[NSAttributedString alloc] initWithString:minValue attributes:valueAttributes];
+            NSAttributedString* attributedMaxValue = [[NSAttributedString alloc] initWithString:maxValue attributes:valueAttributes];
+            
+            args = @[attributedTitle, attributedMinValue, attributedMaxValue];
+            format = NSLocalizedString(@"alarm.thermostat.attribution.format", nil);
+            break;
+        }
+    }
+    
+    return [[NSMutableAttributedString alloc] initWithFormat:format args:args baseColor:[UIColor grey4] baseFont:[UIFont h7]];
+    
+}
+
+- (UIImage*)iconForExpansion:(SENAlarmExpansion*)expansion {
+    switch ([expansion type]) {
+        case SENExpansionTypeThermostat:
+            return [UIImage imageNamed:@"alarmThermostatIcon"];
+        case SENExpansionTypeLights:
+            return [UIImage imageNamed:@"alarmLightIcon"];
+        default:
+            return nil;
     }
 }
 
@@ -314,25 +373,23 @@ static NSString *const HEMAlarmListTimeKey = @"alarms.alarm.meridiem.%@";
     
     NSArray* alarms = [[self alarmService] alarms];
     SENAlarm *alarm = alarms[indexPath.item];
-    SENAlarmExpansion* alarmExpansion = [[self alarmService] lightExpansionIn:alarm];
-    
-    NSString* unit = NSLocalizedString(@"measurement.percentage.unit", nil);
-    NSString* attributionFormat = NSLocalizedString(@"alarm.light.attribution.format", nil);
-    NSString* attribution = [NSString stringWithFormat:attributionFormat,
-                             [alarmExpansion targetRange].max,
-                             unit];
-    [[cell expansionLabel] setText:attribution];
-    [[cell expansionLabel] setFont:[UIFont h7]];
-    [[cell expansionLabel] setTextColor:[UIColor grey6]];
-    [[cell expansionSeparator] setBackgroundColor:[[UIColor grey5] colorWithAlphaComponent:0.1f]];
-    [[cell expansionIconView] setImage:[UIImage imageNamed:@"alarmLightIcon"]];
-    
+
+    for (SENAlarmExpansion* expansion in [alarm expansions]) {
+        if ([expansion isEnable]) {
+            [cell showExpansionWithIcon:[self iconForExpansion:expansion]
+                                   text:[self attributedExpansionTextFor:expansion
+                                                                forAlarm:alarm]
+                                   tyep:[expansion type]];
+        }
+    }
+
     [[cell timeLabel] setTextColor:[UIColor grey5]];
     [[cell meridiemLabel] setTextColor:[UIColor grey5]];
     
     [self updateSwitchInCell:cell forAlarm:alarm atIndexPath:indexPath];
     [self updateDetailTextInCell:cell fromAlarm:alarm];
     [self updateTimeTextInCell:cell fromAlarm:alarm];
+    
     return cell;
 }
 
@@ -369,7 +426,10 @@ static NSString *const HEMAlarmListTimeKey = @"alarms.alarm.meridiem.%@";
 }
 
 - (void)updateTimeTextInCell:(HEMAlarmListCell *)cell fromAlarm:(SENAlarm *)alarm {
+    UIColor* textColor = [alarm isOn] ? [UIColor grey6] : [UIColor grey4];
     cell.timeLabel.text = [self localizedTimeForAlarm:alarm];
+    cell.timeLabel.textColor = textColor;
+    cell.meridiemLabel.textColor = textColor;
     if (![[self alarmService] useMilitaryTimeFormat]) {
         NSString *meridiem = alarm.hour < 12 ? @"am" : @"pm";
         NSString *key = [NSString stringWithFormat:HEMAlarmListTimeKey, meridiem];
@@ -443,16 +503,16 @@ static NSString *const HEMAlarmListTimeKey = @"alarms.alarm.meridiem.%@";
     static CGFloat const HEMAlarmListEmptyCellWidthInset = 32.f;
     UICollectionViewFlowLayout *layout = (id)collectionViewLayout;
     CGFloat width = layout.itemSize.width;
-    
     NSInteger alarmCount = [[[self alarmService] alarms] count];
+    
     if ([self loadError]) {
         return CGSizeMake(width, HEMAlarmListCellHeight);
-    } else if (alarmCount > 0) {
+    } else if ([[[self alarmService] alarms] count] > 0) {
         SENAlarm* alarm = [[self alarmService] alarms][[indexPath row]];
-        CGFloat height = [[self alarmService] hasLightsEnabledForAlarm:alarm]
-            ? HEMAlarmExpansionListCellHeight
-            : HEMAlarmListCellHeight;
-        return CGSizeMake(width, height);
+        CGFloat height = HEMAlarmListCellHeight;
+        NSInteger expansions = [[self alarmService] numberOfEnabledExpansionsIn:alarm];
+        CGFloat expansionHeight = expansions * kHEMAlarmExpansionViewHeight;
+        return CGSizeMake(width, height + expansionHeight);
     } else if (alarmCount == 0) {
         NSAttributedString* attributedText = [self attributedNoAlarmText];
         CGFloat maxWidth = width - (HEMAlarmNoAlarmHorzMargin * 2);
@@ -514,6 +574,7 @@ static NSString *const HEMAlarmListTimeKey = @"alarms.alarm.meridiem.%@";
 
 - (void)toggleEnableSwitch:(UISwitch *)sender {
     NSArray* alarms = [[self alarmService] alarms];
+    NSIndexPath* indexPath = [NSIndexPath indexPathForItem:sender.tag inSection:0];
     __block SENAlarm *alarm = [alarms objectAtIndex:sender.tag];
     BOOL on = [sender isOn];
     
@@ -546,6 +607,7 @@ static NSString *const HEMAlarmListTimeKey = @"alarms.alarm.meridiem.%@";
                                               message:[error localizedDescription]
                                         fromPresenter:strongSelf];
         } else {
+            [[strongSelf collectionView] reloadItemsAtIndexPaths:@[indexPath]];
             [SENAnalytics trackAlarmToggle:alarm];
         }
     }];